controller/pyproject.toml
[project]
name = "controller"
version = "0.1.0"
description = "Add your description here"
authors = [ ]
requires-python = ">=3.11"
dependencies = [
    "cobs>=1.2.2",
    "pyserial>=3.5",
]

[project.optional-dependencies]
non-pi = [
    "opencv-python>=4.12.0.88",
]

[tool.pyright]
typeCheckingMode = "off"
